Cambio en el modelo
Los precios y totales ahora son floats

+++ src/main/java/app/clases/DetalleReserva.java
@@ -3,20 +3,19 @@ package app.clases;
 
 public class DetalleReserva {
 
-	public DetalleReserva(long id, long id_producto, int cantidad, int precio_por_unidad, boolean activo, ProductoProductor producto) {
+	public DetalleReserva(long id, long id_producto, int cantidad, float precio_por_unidad, boolean activo, ProductoProductor producto) {
 		this.setId_reserva(id);
 		this.setId_producto(id_producto);
 		this.setCantidad(cantidad);
 		this.setPrecio_por_unidad(precio_por_unidad);
 		this.setActivo(activo);
 		this.setProducto(producto);
-		
 	}
 
 	private long id_reserva;
 	private long id_producto;
 	private int cantidad;
-	private int precio_por_unidad;
+	private float precio_por_unidad;
 	private boolean activo;
 	private ProductoProductor producto;
 
@@ -52,11 +51,11 @@ public class DetalleReserva {
 		this.activo = activo;
 	}
 	
-	public int getPrecio_por_unidad() {
+	public float getPrecio_por_unidad() {
 		return precio_por_unidad;
 	}
 
-	public void setPrecio_por_unidad(int precio_por_unidad) {
+	public void setPrecio_por_unidad(float precio_por_unidad) {
 		this.precio_por_unidad = precio_por_unidad;
 	}
